+++ lib/src/small_flat_button.dart
@@ -1,0 +1,45 @@
+import 'package:flutter/material.dart';
+
+class SmallFlatButton extends StatelessWidget {
+  final Function onTap;
+  final Widget child;
+  final EdgeInsets margin;
+  final EdgeInsets padding;
+  final Color backgroundColor;
+  final double borderRadius;
+  final double elevation;
+
+  ///A FlatButton with respectively small margin and shape
+  const SmallFlatButton({
+    Key key,
+    @required this.onTap,
+    @required this.child,
+    this.margin = const EdgeInsets.all(8),
+    this.padding = const EdgeInsets.all(8),
+    this.borderRadius = 8,
+    this.backgroundColor = Colors.transparent,
+    this.elevation = 0.0,
+  }) : super(key: key);
+  @override
+  Widget build(BuildContext context) {
+    final shape = RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(borderRadius));
+    return Card(
+      shape: shape,
+      color: backgroundColor,
+      elevation: elevation,
+      margin: margin,
+      child: InkWell(
+        onTap: onTap,
+        customBorder: shape,
+        child: Padding(
+          padding: padding,
+          child: DefaultTextStyle.merge(
+            style: TextStyle(color: Theme.of(context).buttonColor),
+            child: child,
+          ),
+        ),
+      ),
+    );
+  }
+}